Drop minibatches (for now)
Theano auto-differentiation does not appear to be capable of computing
the desired gradient.

We want to take derivatives with respect to a minibatch of parameters,
but define the objective function with respect to all of them. Theano
requires the objective function to be defined only in terms of the
minibatch.

Actually looking at only the minibatch is too noisy and does not
converge.