Implement sample minibatches
This is easy to do, but it also means we cannot trace ELBO now (since it
depends on the minibatch). This was the source of a tricky bug where
Theano complained that an input was not specified in the computation
graph.

Rather than taking balanced subsamples, we take a sliding window over a
permutation, which is balanced in expectation.